--- app/components/UI/Predict/hooks/usePreviousValue.ts
@@ -1,11 +1,0 @@
-import { useEffect, useRef } from 'react';
-
-export function usePreviousValue<T>(value: T): T | undefined {
-  const ref = useRef<T | undefined>(undefined);
-
-  useEffect(() => {
-    ref.current = value;
-  });
-
-  return ref.current;
-}
