+++ app/components/UI/Predict/views/PredictBuyWithAnyToken/hooks/usePredictPayWithAnyTokenTracking.ts
@@ -1,8 +1,8 @@
 import { TransactionStatus } from '@metamask/transaction-controller';
 import { useEffect, useRef } from 'react';
 import { useSelector } from 'react-redux';
-import { RootState } from '../../../../reducers';
-import { selectTransactionMetadataById } from '../../../../selectors/transactionController';
+import { RootState } from '../../../../../../reducers';
+import { selectTransactionMetadataById } from '../../../../../../selectors/transactionController';
 
 interface UsePredictPayWithAnyTokenTrackingParams {
   transactionId?: string;
